Use DeleteFileRequest class When user wants to delete a file the server must check not just the id but all other attributes. If mismatch then do not delete the file In case of file check: id, size, creationTime, lastModifiedTime In case of directory check: id, size, creationTime since we do not sync lastModifiedTime

+++ src/main/java/com/github/tornaia/sync/server/service/FileCommandService.java
@@ -82,12 +82,22 @@ public class FileCommandService {
         }
     }
 
-    public void deleteFile(String clientid, String id) {
+    public void deleteFile(String clientid, String id, long size, long creationDateTime, long modificationDateTime) {
         File file = fileRepository.findOne(id);
         if (file == null) {
-            LOG.info("DELETE not found file: " + id);
+            LOG.info("DELETE File not found: " + id);
             throw new FileNotFoundException(id);
         }
+
+        if (file.getSize() != size || file.getCreationDate() != creationDateTime) {
+            LOG.info("DELETE File attributes mismatch: " + file + ", vs: " + size + ", " + creationDateTime + ", " + modificationDateTime);
+            throw new FileNotFoundException(id);
+        }
+        if (file.isFile() && file.getLastModifiedDate() != modificationDateTime) {
+            LOG.info("DELETE File attributes mismatch: " + file + ", vs: " + size + ", " + creationDateTime + ", " + modificationDateTime);
+            throw new FileNotFoundException(id);
+        }
+
         String path = file.getPath();
         fileRepository.delete(file);
         FileMetaInfo deletedFileMetaInfo = fileToFileMetaInfoConverter.convert(file);
